+++ BattleNetwork/bnArtifact.h
@@ -11,8 +11,8 @@ public:
   Artifact(Field* _field);
   virtual ~Artifact();
 
-  virtual void OnUpdate(float _elapsed) = 0;
+  virtual void OnUpdate(double _elapsed) = 0;
   virtual void OnDelete() = 0;
-  virtual void Update(float _elapsed) final;
+  virtual void Update(double _elapsed) final;
   virtual void AdoptTile(Battle::Tile* tile) final;
 };
