+++ BattleNetwork/bnCanodumbCursor.h
@@ -17,9 +17,9 @@ class CanodumbCursor : public Artifact
 private:
   CanodumbIdleState* parentState; /*!< The context of the Canodumb who spawned it */
   Entity* target; /*!< The enemy to track */
-  float elapsedTime;
-  float movecooldown; /*!< Time remaining between movement */
-  float maxcooldown; /*!< Total time between movement */
+  double elapsedTime;
+  double movecooldown; /*!< Time remaining between movement */
+  double maxcooldown; /*!< Total time between movement */
   Direction direction; /*!< Direction to move */
 
   // Frame select through animation system
@@ -32,7 +32,7 @@ public:
    * @brief If tile is same as target, tells cannon to attack. Otherwise tries to move
    * @param _elapsed
    */
-  void OnUpdate(float _elapsed) override;
+  void OnUpdate(double _elapsed) override;
 
   /**
   * @brief Removes cursor from play
