+++ BattleNetwork/bnExplosionSpriteNode.h
@@ -38,7 +38,7 @@ public:
    * @brief If root increment count is size of numOfExplosions, delete and stop the chain 
    * @param _elapsed in seconds
    */
-  void Update(float _elapsed);
+  void Update(double _elapsed);
 
   /**
    * @brief Used by root. Increment the number of explosions
